+++ lib/async_wrapper.dart
@@ -1,5 +1,5 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+/// A lightweight Flutter package for handling async operations with built-in
+/// loading, success, and error states.
+library async_wrapper;
+
+export 'src/async_wrapper_base.dart';